Main/requirements.txt
confluent_kafka
mysql-connector-python
requests
